API_PROFILES/requirements.txt
fastapi==0.111.0
uvicorn==0.15.0
requests
sqlalchemy
mysql-connector-python
nats-py
